Added Auth layer and extensive testing

+++ packages/citynexus/CityNexus/src/Http/AdminController.php
@@ -25,6 +25,7 @@ class AdminController extends Controller
 
     public function getIndex()
     {
+        $this->authorize('admin', 'view');
         $tables = DB::table('information_schema.tables')->where('table_schema', 'public')->get();
         return view('citynexus::admin.index', compact('tables'));
 
@@ -32,6 +33,8 @@ class AdminController extends Controller
 
     public function getRefreshGeocoding()
     {
+        $this->authorize('admin', 'view');
+
         $properties = Property::whereNull('lat')->get();
 
         foreach($properties as $i)
@@ -42,6 +45,8 @@ class AdminController extends Controller
 
     public function getEditTable(Request $request)
     {
+        $this->authorize('admin', 'view');
+
         $table_name = $request->get('table_name');
 
         return redirect('/tabler/show-table/' . $table_name);
@@ -49,6 +54,8 @@ class AdminController extends Controller
 
     public function getRemoveData(Request $request)
     {
+        $this->authorize('admin', 'delete');
+
         DB::table($request->get('table_name'))->where('id', $request->get('row_id'))->delete();
 
         Session::flash('flash_info', "Row successfully remove");
@@ -59,6 +66,8 @@ class AdminController extends Controller
 
     public function getClearTable($table_name, $remove = false)
     {
+        $this->authorize('admin', 'delete');
+
         if($remove)
         {
             Schema::drop($table_name);
@@ -73,14 +82,10 @@ class AdminController extends Controller
         return redirect('/');
     }
 
-    public function getRemoveChelsea()
-    {
-
-        DB::table('citynexus_properties')->where('city', 'chelsea')->update(['city' => env('CITY')]);
-    }
-
     public function getMergeProperties()
     {
+        $this->authorize('admin', 'view');
+
         $properties = Property::all();
 
         $sorted = array();
